requirements.txt
django==3.2.13
djangorestframework==3.12
djangorestframework-simplejwt==5.1.0
django-cors-headers==3.12.0
django-extensions==3.0.7
psycopg2-binary==2.8.5
six==1.15.0
shortuuid==1.0.0